Create contrib module with several useful traits
