--- smart_contracts/contracts/vm2/vm2-cep18/src/messages.rs
@@ -1,15 +1,0 @@
-use casper_sdk::prelude::*;
-
-#[casper(message)]
-pub struct Transfer {
-    pub from: Option<Entity>,
-    pub to: Entity,
-    pub amount: u64,
-}
-
-#[casper(message)]
-pub struct Approve {
-    pub owner: Entity,
-    pub spender: Entity,
-    pub amount: u64,
-}
