+++ smart_contracts/sdk/src/contrib.rs
@@ -1,1 +1,4 @@
 pub mod access_control;
+pub mod cep18;
+pub mod ownable;
+pub mod pausable;
