--- smart_contracts/sdk/tests/test_contrib.rs
@@ -1,4 +1,0 @@
-use casper_sdk::enumerate_AccessControl_symbols;
-
-#[test]
-fn contrib() {}
